Modifications to git-svn-clone-externals and git-svn-externals-check.
Made some modifications to git-svn-clone-externals so that it handles
appropriately svn branches with names containing the word `trunk' in them, and
to ignore CDPATH if it is set in the environment.

Made also a small addition to git-svn-externals-check so that one can pass it
options like -s and -uno, as in git status.